+++ packages/wallet-bitcoin/src/bitcoin-core.ts
@@ -20,15 +20,17 @@ class ECCManager {
   // @ts-ignore
   private bitcoinlabEccPair: ECPairAPI
 
+  private ready = false
+
   constructor() {
     // Lazy load
-    this.tinyEcc = import('tiny-secp256k1').then(v => {
+    import('tiny-secp256k1').then(v => {
       console.log('tiny-secp256k1 loaded')
       this.tinyEcc = v
       // @ts-ignore
       this.tinyEccPair = (ECPairModule.default ? ECPairModule.default : ECPairModule)(this.tinyEcc)
     })
-    this.bitcoinlabEcc = import('@bitcoinerlab/secp256k1').then(v => {
+    import('@bitcoinerlab/secp256k1').then(v => {
       console.log('@bitcoinerlab/secp256k1 loaded')
       this.bitcoinlabEcc = v
       // @ts-ignore
@@ -36,6 +38,46 @@ class ECCManager {
         this.bitcoinlabEcc
       )
     })
+
+    this.waitForAllEccReady().then(() => {
+      this.setEccType(this.eccType)
+      this.ready = true
+    })
+  }
+
+  private waitForAllEccReady(): Promise<void> {
+    return new Promise((resolve, reject) => {
+      const check = () => {
+        if (this.tinyEcc && this.bitcoinlabEcc) {
+          resolve()
+        } else {
+          setTimeout(check, 50)
+        }
+      }
+      check()
+    })
+  }
+
+  waitForReady(): Promise<void> {
+    return new Promise((resolve, reject) => {
+      const check = () => {
+        if (this.ready) {
+          resolve()
+        } else {
+          setTimeout(check, 50)
+        }
+      }
+      check()
+    })
+  }
+
+  setEccType(eccType: EccType) {
+    this.eccType = eccType
+    if (this.eccType === EccType.tiny) {
+      bitcoin.initEccLib(this.tinyEcc)
+    } else {
+      bitcoin.initEccLib(this.bitcoinlabEcc)
+    }
   }
 
   get ecc() {