+++ packages/wallet-bitcoin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @unisat/wallet-bitcoin
 
+## 0.1.13
+
+### Patch Changes
+
+- fix
+
 ## 0.1.12
 
 ### Patch Changes